requirements.txt
PySide6
pdfplumber
easyocr
python-docx
pyttsx3
fpdf
transformers
langid
torch